+++ data/slack-jackson-dto-test-extensions/src/main/kotlin/com/kreait/slack/api/contract/jackson/common/types/IdentityExtension.kt
@@ -2,6 +2,6 @@ package com.kreait.slack.api.contract.jackson.common.types
 
 fun Identity.User.Companion.sample() = Identity.User("", "", "", "", "", "", "", "")
 
-fun Identity.Team.Companion.sample() = Identity.Team("", "")
+fun Team.Companion.sample() = Team("", "")
 
-fun Identity.Companion.sample() = Identity(Identity.User.sample(), Identity.Team.sample())
+fun Identity.Companion.sample() = Identity(Identity.User.sample(), Team.sample())