+++ data/slack-jackson-dto-test-extensions/src/main/kotlin/com/kreait/slack/api/contract/jackson/group/users/UsersIdentityExtensions.kt
@@ -1,10 +1,11 @@
 package com.kreait.slack.api.contract.jackson.group.users
 
 import com.kreait.slack.api.contract.jackson.common.types.Identity
+import com.kreait.slack.api.contract.jackson.common.types.Team
 import com.kreait.slack.api.contract.jackson.common.types.sample
 
 fun SuccessfulIdentityResponse.Companion.sample(): SuccessfulIdentityResponse {
-    return SuccessfulIdentityResponse(true, Identity(Identity.User.sample(), Identity.Team.sample()))
+    return SuccessfulIdentityResponse(true, Identity(Identity.User.sample(), Team.sample()))
 }
 
 fun ErrorIdentityResponse.Companion.sample(): ErrorIdentityResponse {