+++ data/slack-jackson-dto/src/main/kotlin/com/kreait/slack/api/contract/jackson/group/oauth/Access.kt
@@ -3,6 +3,7 @@ package com.kreait.slack.api.contract.jackson.group.oauth
 import com.fasterxml.jackson.annotation.JsonProperty
 import com.fasterxml.jackson.annotation.JsonSubTypes
 import com.fasterxml.jackson.annotation.JsonTypeInfo
+import com.kreait.slack.api.contract.jackson.common.types.Team
 import com.kreait.slack.api.contract.jackson.util.JacksonDataClass
 
 
@@ -13,17 +14,18 @@ data class AccessRequest(val clientId: String, val client_secret: String, val co
 
 @JsonTypeInfo(use = JsonTypeInfo.Id.NAME, include = JsonTypeInfo.As.PROPERTY, property = "ok", visible = true)
 @JsonSubTypes(
-        JsonSubTypes.Type(value = SuccessfullAccessResponse::class, name = "true"),
-        JsonSubTypes.Type(value = ErrorAccessResponse::class, name = "false")
+    JsonSubTypes.Type(value = SuccessfullAccessResponse::class, name = "true"),
+    JsonSubTypes.Type(value = ErrorAccessResponse::class, name = "false")
 )
 @JacksonDataClass
 sealed class AccessResponse(@JsonProperty("ok") open val ok: Boolean)
 
 
 @JacksonDataClass
-data class ErrorAccessResponse constructor(override val ok: Boolean,
-                                           @JsonProperty("error") val error: String)
-    : AccessResponse(ok) {
+data class ErrorAccessResponse constructor(
+    override val ok: Boolean,
+    @JsonProperty("error") val error: String
+) : AccessResponse(ok) {
     companion object
 }
 
@@ -40,45 +42,31 @@ data class ErrorAccessResponse constructor(override val ok: Boolean,
  * @property bot the bot object with the according information
  */
 data class SuccessfullAccessResponse(
-        @JsonProperty("ok") override val ok: Boolean,
-        @JsonProperty("access_token") val accessToken: String,
-        @JsonProperty("scope") val scope: String,
-        @JsonProperty("user_id") val userId: String,
-        @JsonProperty("team_name") val teamName: String,
-        @JsonProperty("team_id") val teamId: String,
-        @JsonProperty("incoming_webhook") val incomingWebhook: IncomingWebhook,
-        @JsonProperty("bot") val bot: Bot
+    @JsonProperty("ok") override val ok: Boolean,
+    @JsonProperty("access_token") val accessToken: String,
+    @JsonProperty("token_type") val tokenType: String,
+    @JsonProperty("scope") val scope: String,
+    @JsonProperty("bot_user_id") val botUserId: String,
+    @JsonProperty("app_id") val appId: String,
+    @JsonProperty("team") val team: Team,
+    @JsonProperty("enterprise") val enterprise: Enterprise?
 ) : AccessResponse(ok) {
-    companion object
-}
+    companion object {}
 
-/**
- * The bot property which is returned when you request the bot scope
- *
- * @property botUserId the bot user id of your bot
- * @property botAccessToken the accesstoken of your bot
- */
-data class Bot(
-        @JsonProperty("bot_user_id") val botUserId: String,
-        @JsonProperty("bot_access_token") val botAccessToken: String
-) {
-    companion object
-}
+    data class Enterprise(
+        @JsonProperty("name") val name: String?,
+        @JsonProperty("id") val id: String
+    ) {
+        companion object
+    }
 
-/**
- * Incoming webhook which is only included if incoming_webhook scope is requested
- *
- * @property channel the channel name of the webhook
- * @property channelId the channel id of the webhook
- * @property configurationUrl the configuration page url of that webhook
- * @property url the url where you can post requests to e.g. to send a message
- */
-data class IncomingWebhook(
-        @JsonProperty("channel") val channel: String,
-        @JsonProperty("channel_id") val channelId: String,
-        @JsonProperty("configuration_url") val configurationUrl: String,
-        @JsonProperty("url") val url: String
-) {
-    companion object
+    data class AuthedUser(
+        @JsonProperty("id") val id: String,
+        @JsonProperty("scope") val scope: String,
+        @JsonProperty("accessToken") val accessToken: String,
+        @JsonProperty("tokenType") val tokenType: String
+    ) {
+        companion object
+    }
 }