+++ starter/slack-spring-boot/src/main/kotlin/com/kreait/slack/broker/store/team/TeamStore.kt
@@ -34,29 +34,20 @@ class TeamNotFoundException(override val message: String?) : RuntimeException(me
 /**
  * describes a Team that installed your App
  */
-data class Team(val teamId: String,
-                val teamName: String,
-                val incomingWebhook: IncomingWebhook?,
-                val bot: Bot) {
+data class Team(
+    val teamId: String,
+    val teamName: String?,
+    val bot: Bot
+) {
     companion object {}
 
     /**
      * contains relevant information about the installed bot
      */
     data class Bot(
-            val userId: String,
-            val accessToken: String) {
-        companion object
-    }
-
-    /**
-     * Incoming-webhook object that can be used to post messages to a channel without user-interaction
-     */
-    data class IncomingWebhook(
-            val channel: String,
-            val channelId: String,
-            val configurationUrl: String,
-            val url: String) {
+        val userId: String,
+        val accessToken: String
+    ) {
         companion object
     }
 }