+++ starter/slack-spring-boot/src/test/kotlin/com/kreait/slack/broker/extensions/TeamStoreTestExtension.kt
@@ -2,8 +2,6 @@ package com.kreait.slack.broker.extensions
 
 import com.kreait.slack.broker.store.team.Team
 
-fun Team.Companion.sample(): Team = Team("TestTeamId", "TestTeamName", Team.IncomingWebhook.sample(), Team.Bot.sample())
-
-fun Team.IncomingWebhook.Companion.sample(): Team.IncomingWebhook = Team.IncomingWebhook("TestChannel", "TestChannelId", "http://configuration-test.com", "http://test.com")
+fun Team.Companion.sample(): Team = Team("TestTeamId", "TestTeamName", Team.Bot.sample())
 
 fun Team.Bot.Companion.sample(): Team.Bot = Team.Bot("UserId", "acccessToken")